refactor(BA-6483): split AppConfigPolicy into admin/non-admin service & processors
Mirror the login_client_type convention: non-admin Service/Processors (get,
scoped_search) and AdminService/AdminProcessors (admin_search, bulk
create/update/purge), wired as app_config_policy + app_config_policy_admin.
